requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
h11==0.12.0
idna==3.3
MouseInfo==0.1.3
outcome==1.1.0
Pillow==8.4.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-telegram-bot==13.8.1
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Wand==0.6.7
wsproto==1.0.0
